api/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0

psycopg2==2.9.10

fabric==3.2.2

pytz==2025.1

uwsgi==2.0.28

boto3==1.37.33
botocore==1.37.33
django-storages==1.14.6

python-dotenv==1.1.0
requests==2.32.3

celery==5.5.0
redis==5.2.1
flower==2.0.1
django-redis==5.4.0

daphne==4.2.1
channels==4.2.2
channels-redis==4.2.1

sendgrid==6.12.2
numpy==2.2.6
pandas==2.3.1
openai==1.82.1
pgvector==0.4.1
pdf2image==1.17.0
Pillow==11.3.0
opencv-python==4.12.0.88
opencv-contrib-python==4.12.0.88
pypandoc==1.15
weasyprint==66.0
google-cloud-vision==3.10.2
google-cloud-documentai==3.5.0
google-cloud-texttospeech==2.27.0
google-cloud-speech==2.33.0
google-generativeai==0.8.5
beautifulsoup4==4.13.4
tiktoken==0.11.0
mutagen==1.47.0
PyPDF2==3.0.1
azure-cognitiveservices-speech==1.45.0
pydub==0.25.1
twilio==9.7.0
beautifulsoup4==4.13.4
tiktoken==0.11.0
mutagen==1.47.0